+++ opsgenie/provider.go
@@ -24,6 +24,7 @@ func Provider() terraform.ResourceProvider {
 
 		ResourcesMap: map[string]*schema.Resource{
 			"opsgenie_team":              resourceOpsGenieTeam(),
+			"opsgenie_team_routing_rule": resourceOpsGenieTeamRoutingRule(),
 			"opsgenie_user":              resourceOpsGenieUser(),
 			"opsgenie_user_contact":      resourceOpsGenieUserContact(),
 			"opsgenie_escalation":        resourceOpsgenieEscalation(),