+++ CHANGELOG.md
@@ -1,4 +1,16 @@
 ## 0.2.6 (Unreleased)
+
+IMPROVEMENTS:
+
+* **Updated opsgenie-go-sdk-v2** 
+
+* **New Resource:** Team Routing Rule
+
+BUGFIXES:
+
+* 'Random' NotifyType added.
+
+
 ## 0.2.5 (November 19, 2019)
 
 IMPROVEMENTS: